encryptString done. (all major components done)
I'd like to find some original ciphers and decrypts
with the wheel settings included. Then I could see
that this thing is actually working properly.

As is, the wheel settings are obviously bogus, as
are the initial position settings.

Also, I should check whether the wheels step before
or after each character is encoded.